+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## UNRELEASED
+### Added
+- Throttled listeners can now be canceled (useful for cleanup):
+  ```js
+  var throttledListener = throttle(listener);
+  window.addEventListener('resize', throttledListener);
+  submitButton.addEventListener('click', () => {
+    window.removeEventListener('resize', throttledListener);
+    // prevent any queued calls from executing on the next animation frame:
+    throttledListener.cancel();
+  })
+  ```
+
+### Changed
+- Updated types to use generics; `throttle` will now return a function
+  of the same type it was passed.
+- frame-throttle now runs in strict mode
+
 ### Fixed
 - Binding a throttled listener with `.bind()` resulted in both the bound and
   unbound listeners being throttled together.
@@ -18,10 +35,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   boundThrottledListener1();
   boundThrottledListener2();
   ```
-
-### Changed
-- Updated types to use generics; `throttle` will now return a function
-  of the same type it was passed.
 
 ## [2.0.1] - 2016-09-26
 ### Fixed